+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 #Changelog#
-v4.0.1
+v4.1.0
 - Added support for config files one level above current working directory
+- Added Environment class to core library
+- Moved database connection to init earlier in the application layer
 
 v4.0
 - Added feature for language/i18n, enabled in config files